Move format field to parent class
